requirements/dev.txt
# Dependencies for the dev environment
-r base.txt

# test tools
django-nose>=1.4.2
coverage>=6

# debug tools
django-debug-toolbar>=3.2
